+++ apps/storybook/.storybook/orderlyProvider.tsx
@@ -41,6 +41,7 @@ import {
 } from "@orderly.network/wallet-connector-privy";
 import { CustomProductNav } from "../src/components/customProductNav/indx";
 import { useOrderlyConfig } from "../src/hooks/useOrderlyConfig";
+import { customChainsAbstarct } from "./customChains";
 import { CustomConfigStore } from "./customConfigStore";
 
 const network = WalletAdapterNetwork.Devnet;
@@ -118,12 +119,12 @@ export const OrderlyProvider: FC<{ children: ReactNode }> = (props) => {
       }}
     >
       <WalletConnectorPrivyProvider
-        termsOfUse="https://learn.woo.org/legal/terms-of-use"
+        // termsOfUse="https://learn.woo.org/legal/terms-of-use"
         network={Network.testnet}
         headerProps={{
           mobile: <CustomProductNav />,
         }}
-        // customChains={customerChains}
+        // customChains={customChainsAbstarct}
         privyConfig={{
           appid: "cm50h5kjc011111gdn7i8cd2k",
           config: {
@@ -158,14 +159,14 @@ export const OrderlyProvider: FC<{ children: ReactNode }> = (props) => {
             console.log("-- error", error, adapter);
           },
         }}
-        // abstractConfig={{}}
+        abstractConfig={{}}
       >
         <OrderlyAppProvider
-          customChains={customChains as any}
+          // customChains={customChains as any}
           configStore={configStore}
           appIcons={config.orderlyAppProvider.appIcons}
           restrictedInfo={config.orderlyAppProvider.restrictedInfo}
-          // customChains={customerChains}
+          // customChains={customChainsAbstarct}
           // defaultChain={{testnet: customChains.testnet[0], mainnet: customChains.mainnet[0]}}
         >
           {props.children}