+++ packages/component/src/block/accountStatus/desktop/leverageDialog.tsx
@@ -1,4 +1,7 @@
 import { FC, PropsWithChildren, useEffect, useRef, useState } from "react";
+import { useLeverage } from "@orderly.network/hooks";
+import { useMarginRatio } from "@orderly.network/hooks";
+import Button from "@/button";
 import {
   Dialog,
   DialogBody,
@@ -8,21 +11,16 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/dialog";
-import { LeverageEditor } from "../sections/leverageEditor";
-import Button from "@/button";
-import { useLeverage } from "@orderly.network/hooks";
-import { toast } from "@/toast";
-import { useMarginRatio } from "@orderly.network/hooks";
 import { Numeral } from "@/text/numeral";
-
-interface Props {}
+import { toast } from "@/toast";
+import { LeverageEditor } from "../sections/leverageEditor";
 
 export const LeverageDialog: FC<PropsWithChildren> = (props) => {
   const [open, setOpen] = useState(false);
   const { currentLeverage } = useMarginRatio();
 
-  const [maxLeverage, { update, config: leverageLevers, isMutating }] =
-    useLeverage();
+  const { update, maxLeverage, isLoading, leverageLevers } = useLeverage();
+
   const nextLeverage = useRef(maxLeverage ?? 0);
 
   useEffect(() => {
@@ -45,7 +43,7 @@ export const LeverageDialog: FC<PropsWithChildren> = (props) => {
       (err: Error) => {
         console.dir(err);
         toast.error(err.message);
-      }
+      },
     );
   };
 
@@ -62,7 +60,7 @@ export const LeverageDialog: FC<PropsWithChildren> = (props) => {
             <div className="orderly-flex orderly-gap-1">
               <span>Current:</span>
               <Numeral className="orderly-text-base-contrast" surfix={"x"}>
-                {currentLeverage}
+                {currentLeverage!}
               </Numeral>
             </div>
           </div>
@@ -89,7 +87,7 @@ export const LeverageDialog: FC<PropsWithChildren> = (props) => {
             id="orderly-desktop-leverage-dialog-save"
             fullWidth
             onClick={() => onSubmit()}
-            loading={isMutating}
+            loading={isLoading}
           >
             Save
           </Button>