+++ packages/core/src/assets.ts
@@ -1,4 +1,5 @@
 import {
+  ABSTRACT_CHAIN_ID_MAP,
   ABSTRACT_TESTNET_CHAINID,
   API,
   ApiError,
@@ -59,20 +60,26 @@ export class Assets {
     // const EIP_712signatured = await this.account.signTypedData(
     //   toSignatureMessage
     // );
+    const messageData = {
+      receiver: this.account.stateValue.address,
+      token,
+      brokerId: this.configStore.get("brokerId"),
+      amount: this.account.walletAdapter.parseUnits(amount),
+      nonce,
+      timestamp,
+      // domain,
+      verifyContract:
+        this.contractManger.getContractInfoByEnv().verifyContractAddress,
+    };
+
+    const agwGobalAddress = this.account.getAdditionalInfo()?.AGWAddress ?? "";
+
+    if (ABSTRACT_CHAIN_ID_MAP.has(chainId) && agwGobalAddress) {
+      messageData.receiver = agwGobalAddress;
+    }
 
     const { message, signatured, domain } =
-      await this.account.walletAdapter.generateWithdrawMessage({
-        // chainId,
-        receiver: this.account.stateValue.address!,
-        token,
-        brokerId: this.configStore.get("brokerId"),
-        amount: this.account.walletAdapter.parseUnits(amount),
-        nonce,
-        timestamp,
-        // domain,
-        verifyContract:
-          this.contractManger.getContractInfoByEnv().verifyContractAddress,
-      });
+      await this.account.walletAdapter.generateWithdrawMessage(messageData);
 
     const data = {
       signature: signatured,
@@ -166,9 +173,15 @@ export class Assets {
 
     let usdcAddress = address ?? contractAddress.usdcAddress;
     let userAddress = this.account.stateValue.address;
-    if (this.account.walletAdapter.chainId === ABSTRACT_TESTNET_CHAINID) {
-      usdcAddress = contractAddress.abstractTestnetUSDCAddress ?? "";
-      userAddress = this.account.getAdditionalInfo()?.AGWAddress ?? "";
+    if (ABSTRACT_CHAIN_ID_MAP.has(this.account.walletAdapter.chainId)) {
+      usdcAddress = contractAddress.abstractUSDCAddress ?? "";
+    }
+    const agwGobalAddress = this.account.getAdditionalInfo()?.AGWAddress ?? "";
+    if (
+      ABSTRACT_CHAIN_ID_MAP.has(this.account.walletAdapter.chainId) &&
+      agwGobalAddress
+    ) {
+      userAddress = agwGobalAddress;
     }
     const result = await this.account.walletAdapter?.call(
       usdcAddress,
@@ -240,12 +253,18 @@ export class Assets {
       tempVaultAddress = contractAddress.monadTestnetVaultAddress ?? "";
       tempUSDCAddress = contractAddress.monadTestnetUSDCAddress;
     }
-    if (this.account.walletAdapter.chainId === ABSTRACT_TESTNET_CHAINID) {
-      tempVaultAddress = contractAddress.abstractTestnetVaultAddress ?? "";
-      tempUSDCAddress = contractAddress.abstractTestnetUSDCAddress ?? "";
-      userAddress = this.account.getAdditionalInfo()?.AGWAddress ?? "";
+    if (ABSTRACT_CHAIN_ID_MAP.has(this.account.walletAdapter.chainId)) {
+      tempVaultAddress = contractAddress.abstractVaultAddress ?? "";
+      tempUSDCAddress = contractAddress.abstractUSDCAddress ?? "";
     }
-    console.log("get allowance", tempUSDCAddress, tempVaultAddress);
+    const agwGobalAddress = this.account.getAdditionalInfo()?.AGWAddress ?? "";
+    if (
+      ABSTRACT_CHAIN_ID_MAP.has(this.account.walletAdapter.chainId) &&
+      agwGobalAddress
+    ) {
+      userAddress = agwGobalAddress;
+    }
+
     const result = await this.account.walletAdapter?.call(
       tempUSDCAddress ?? "",
       "allowance",
@@ -304,9 +323,8 @@ export class Assets {
       tempVaultAddress = contractAddress.monadTestnetVaultAddress ?? "";
       tempUSDCAddress = contractAddress.monadTestnetUSDCAddress ?? "";
     }
-    if (this.account.walletAdapter.chainId === ABSTRACT_TESTNET_CHAINID) {
-      tempVaultAddress = contractAddress.abstractTestnetVaultAddress ?? "";
-      tempUSDCAddress = contractAddress.abstractTestnetUSDCAddress ?? "";
+    if (ABSTRACT_CHAIN_ID_MAP.has(this.account.walletAdapter.chainId)) {
+      tempVaultAddress = contractAddress.abstractVaultAddress ?? "";
     }
 
     const result = await this.account.walletAdapter?.call(
@@ -317,7 +335,6 @@ export class Assets {
         abi: contractAddress.usdcAbi,
       },
     );
-
     return result;
   }
 
@@ -382,10 +399,8 @@ export class Assets {
       vaultAddress = contractAddress.monadTestnetVaultAddress ?? "";
       // depositData["USDCAddress"] = contractAddress.monadTestnetUSDCAddress ?? "";
     }
-    if (chain.chain_id === ABSTRACT_TESTNET_CHAINID) {
-      vaultAddress = contractAddress.abstractTestnetVaultAddress ?? "";
-      // tips:  deposit fee not use global address, juset sign address
-      // userAddress = this.account.getAdditionalInfo()?.AGWAddress ?? "";
+    if (ABSTRACT_CHAIN_ID_MAP.has(this.account.walletAdapter.chainId)) {
+      vaultAddress = contractAddress.abstractVaultAddress ?? "";
     }
 
     console.log("xxx get deposit fee", {
@@ -439,24 +454,40 @@ export class Assets {
     if (this.account.walletAdapter.chainId === MONAD_TESTNET_CHAINID) {
       vaultAddress = contractAddress.monadTestnetVaultAddress ?? "";
     }
-    if (this.account.walletAdapter.chainId === ABSTRACT_TESTNET_CHAINID) {
-      vaultAddress = contractAddress.abstractTestnetVaultAddress ?? "";
-      // userAddress = this.account.getAdditionalInfo()?.AGWAddress ?? "";
+    if (ABSTRACT_CHAIN_ID_MAP.has(this.account.walletAdapter.chainId)) {
+      vaultAddress = contractAddress.abstractVaultAddress ?? "";
+    }
+    const agwGobalAddress = this.account.getAdditionalInfo()?.AGWAddress ?? "";
+
+    let contractMethod = "deposit";
+    let fromAddress = userAddress;
+    let contractData: any = [depositData];
+    console.log(
+      "agw address",
+      agwGobalAddress,
+      this.account.walletAdapter.chainId,
+    );
+    if (
+      ABSTRACT_CHAIN_ID_MAP.has(this.account.walletAdapter.chainId) &&
+      agwGobalAddress
+    ) {
+      contractMethod = "depositTo";
+      fromAddress = agwGobalAddress;
+      contractData = [userAddress, depositData];
     }
     console.log("xxx deposit", {
-      userAddress,
-      vaultAddress,
-      depositData,
+      fromAddress,
+      contractMethod,
+      contractData,
     });
     const result = await this.account.walletAdapter?.sendTransaction(
       vaultAddress,
-      "deposit",
+      contractMethod,
       {
-        from: userAddress!,
+        from: fromAddress!,
         to: vaultAddress,
-        data: [depositData],
+        data: contractData,
         value: fee,
-        // value: BigInt(114557727841647),
       },
       {
         abi: contractAddress.vaultAbi,